requirements.txt
anthropic
python-dotenv
playwright
python-pptx
Pillow
flask
supabase
python-slugify
beautifulsoup4
markdownify
requests
google-genai
boto3
Pillow